backend/requirements.txt
# Workspace version: 1.44.26
databutton==0.38.34
ipykernel==6.29.5
fastapi==0.111.0
python-multipart==0.0.9
uvicorn[standard]==0.29.0

openai
beautifulsoup4
requests
httpx
pydantic-settings
firebase-admin
PyPDF2
bigcommerce
user-agents
qrcode
pillow
svgwrite
aiohttp
pytest